+++ activerecord/test/cases/type/decimal_test.rb
@@ -10,6 +10,11 @@ module ActiveRecord
         assert_equal BigDecimal.new("1"), type.type_cast_from_user(:"1")
       end
 
+      def test_type_cast_decimal_from_float_with_large_precision
+        type = Decimal.new(precision: ::Float::DIG + 2)
+        assert_equal BigDecimal.new("123.0"), type.type_cast_from_user(123.0)
+      end
+
       def test_type_cast_decimal_from_rational_with_precision
         type = Decimal.new(precision: 2)
         assert_equal BigDecimal("0.33"), type.type_cast_from_user(Rational(1, 3))